Move `isPartialApply` out of deck in to GDR (#182)
* Add support for foreign key references in partial sync

The b.intermediate collection is populated when entities
are discovered in targetState, allowing for in-file foreign
key references.

The new `deck gateway apply` command allows you to sync
partial configuration, and so must populate b.intermediate
from the currentState.

This results in duplicate entity errors, and so an
AddIgnoringDuplicates method has been added to entities
that may be a foreign key

* Increase golangci-lint timeout to 10m

* Update tests to use require.NoError()

---------

+++ pkg/diff/diff.go
@@ -585,7 +585,7 @@ func generateDiffString(e crud.Event, isDelete bool, noMaskValues bool) (string,
 func (sc *Syncer) Solve(ctx context.Context, parallelism int, dry bool, isJSONOut bool) (Stats,
 	[]error, EntityChanges,
 ) {
-	// TODO https://github.com/Kong/go-database-reconciler/issues/22
+	// TODO https://github.com/Kong/go-database-reconciler/issues/22/
 	// this can probably be extracted to clients (only deck uses it) by having clients count events through the result
 	// channel, rather than returning them from Solve.
 	stats := Stats{
@@ -600,7 +600,9 @@ func (sc *Syncer) Solve(ctx context.Context, parallelism int, dry bool, isJSONOu
 		case crud.Update:
 			stats.UpdateOps.Increment(1)
 		case crud.Delete:
-			stats.DeleteOps.Increment(1)
+			if !sc.noDeletes {
+				stats.DeleteOps.Increment(1)
+			}
 		}
 	}
 
@@ -727,16 +729,18 @@ func (sc *Syncer) Solve(ctx context.Context, parallelism int, dry bool, isJSONOu
 				}
 			}
 		case crud.Delete:
-			// TODO https://github.com/Kong/go-database-reconciler/issues/22 this currently supports either the entity
-			// actions channel or direct console outputs to allow a phased transition to the channel only. Existing console
-			// prints and JSON blob building will be moved to the deck client.
-			if sc.enableEntityActions {
-				actionResult.Action = DeleteAction
-			} else {
-				if isJSONOut {
-					output.Deleting = append(output.Deleting, item)
+			if !sc.noDeletes {
+				// TODO https://github.com/Kong/go-database-reconciler/issues/22 this currently supports either the entity
+				// actions channel or direct console outputs to allow a phased transition to the channel only. Existing console
+				// prints and JSON blob building will be moved to the deck client.
+				if sc.enableEntityActions {
+					actionResult.Action = DeleteAction
 				} else {
-					sc.deletePrintln("deleting", e.Kind, c.Console())
+					if isJSONOut {
+						output.Deleting = append(output.Deleting, item)
+					} else {
+						sc.deletePrintln("deleting", e.Kind, c.Console())
+					}
 				}
 			}
 		default:
